Updated adjusted reveal css.

+++ _update.py
@@ -28,13 +28,14 @@ def wide_clean():
     # (must be in the root directory)
     thisdir = os.getcwd()
     dirs = [os.path.join('lib', 'doconce', 'docstrings'),
-            os.path.join('doc', 'tutorial'),
-            os.path.join('doc', 'manual'),
-            os.path.join('doc', 'quickref'),
-            os.path.join('doc', 'slides'),
-            os.path.join('doc', 'blog'),
+            os.path.join('doc', 'src', 'tutorial'),
+            os.path.join('doc', 'src', 'manual'),
+            os.path.join('doc', 'src', 'quickref'),
+            os.path.join('doc', 'src', 'slides'),
+            os.path.join('doc', 'src', 'blog'),
+            os.path.join('doc', 'src', 'design'),
             'test',
-            'bundled',
+            #'bundled',  # leave cloned repos and delete them manually
             ]
     for d in dirs:
         os.chdir(d)
@@ -64,48 +65,53 @@ def zipfiles2lib():
 def pack_reveal_deck_csss():
     system('sh clean.sh')
 
-    rmtree('reveal.js')
-    system('git clone git://github.com/hakimel/reveal.js.git')
+    if clone:
+        rmtree('reveal.js')
+        system('git clone git://github.com/hakimel/reveal.js.git')
     os.system('cp doconce_modifications/reveal/css/reveal*.css reveal.js/css/')
     os.system('cp doconce_modifications/reveal/css/theme/*.css reveal.js/css/theme/')
     zip_dir('reveal.js')
 
-    rmtree('csss')
-    system('git clone git://github.com/LeaVerou/csss.git')
+    if clone:
+        rmtree('csss')
+        system('git clone git://github.com/LeaVerou/csss.git')
     os.system('cp doconce_modifications/csss/*.css csss/')
     zip_dir('csss')
 
-    rmtree('deck.js')
-    system('git clone git://github.com/imakewebthings/deck.js.git')
-    rmtree('mnml')
-    system('git clone git://github.com/duijf/mnml.git')
-    system('cp mnml/mnml.css deck.js/themes/style')
-    rmtree('deckjs-theme-mozilla')
-    system('git clone git://github.com/groovecoder/deckjs-theme-mozilla.git')
-    system('cp deckjs-theme-mozilla/*.*css deck.js/themes/style')
-    rmtree('deck.js-codemirror')
-    system('git clone git://github.com/iros/deck.js-codemirror.git')
-    os.mkdir('deck.js/extensions/codemirror')
-    system('cp -r deck.js-codemirror/* deck.js/extensions/codemirror/')
-    rmtree('deck.ext.js')
-    system('git clone git://github.com/barraq/deck.ext.js.git')
-    system('cp -r deck.ext.js/themes/style/*.*css deck.js/themes/style/')
-    rmtree('deck.pointer.js')
-    system('git clone git://github.com/mikeharris100/deck.pointer.js.git')
-    os.mkdir('deck.js/extensions/pointer')
-    system('cp -r deck.pointer.js/deck.pointer.* deck.js/extensions/pointer/')
-    rmtree('presenterview')
-    system('git clone git://github.com/stvnwrgs/presenterview.git')
-    system('cp -r presenterview/ deck.js/extensions/')
-    rmtree('deck.annotate.js')
-    system('git clone git://github.com/nemec/deck.annotate.js.git')
-    system('cp -r deck.annotate.js deck.js/extensions/')
-    rmtree('deck.js-notes')
-    system('git clone git@github.com:freekh/deck.js-notes.git')
-    system('cp -r deck.js-notes deck.js/extensions/notes')
+    if clone:
+        rmtree('deck.js')
+        system('git clone git://github.com/imakewebthings/deck.js.git')
+        rmtree('mnml')
+        system('git clone git://github.com/duijf/mnml.git')
+        rmtree('deckjs-theme-mozilla')
+        system('git clone git://github.com/groovecoder/deckjs-theme-mozilla.git')
+        rmtree('deck.js-codemirror')
+        system('git clone git://github.com/iros/deck.js-codemirror.git')
+        os.mkdir('deck.js/extensions/codemirror')
+        rmtree('deck.ext.js')
+        system('git clone git://github.com/barraq/deck.ext.js.git')
+        rmtree('deck.pointer.js')
+        system('git clone git://github.com/mikeharris100/deck.pointer.js.git')
+        os.mkdir('deck.js/extensions/pointer')
+        rmtree('presenterview')
+        system('git clone git://github.com/stvnwrgs/presenterview.git')
+        rmtree('deck.annotate.js')
+        system('git clone git://github.com/nemec/deck.annotate.js.git')
+        rmtree('deck.js-notes')
+        system('git clone git@github.com:freekh/deck.js-notes.git')
+
+        system('cp mnml/mnml.css deck.js/themes/style')
+        system('cp deckjs-theme-mozilla/*.*css deck.js/themes/style')
+        system('cp -r deck.js-codemirror/* deck.js/extensions/codemirror/')
+        system('cp -r deck.ext.js/themes/style/*.*css deck.js/themes/style/')
+        system('cp -r deck.pointer.js/deck.pointer.* deck.js/extensions/pointer/')
+        system('cp -r presenterview/ deck.js/extensions/')
+        system('cp -r deck.annotate.js deck.js/extensions/')
+        system('cp -r deck.js-notes deck.js/extensions/notes')
 
     system('cp doconce_modifications/deck/core/*.css deck.js/core/')
     system('cp doconce_modifications/deck/themes/style/*.css deck.js/themes/style/')
+
     # this find will always generate errors..., use os.system
     os.system("find deck.js/extensions -name '.git' -exec rm -rf {} \;")
     zip_dir('deck.js')
@@ -130,9 +136,13 @@ if __name__ == '__main__':
     os.system('rm -f *.zip')
 
     if len(sys.argv) == 1:
-        print 'Usage: python _update.py all | local'
+        print 'Usage: python _update.py all | | all-noclone | local'
         sys.exit(1)
     if sys.argv[1] == 'all':
+        clone = True
+        run_all()
+    elif sys.argv[1] == 'all-noclone':
+        clone = False
         run_all()
     else:
         #func = sys.argv[1]
